+++ packages/components/organisms/f-takeawaypay-activation/src/tenants/index.js
@@ -1,7 +1,0 @@
-import uk from './en-GB';
-
-const tenantConfigs = {
-    'en-GB': uk
-};
-
-export default tenantConfigs;