+++ packages/components/pages/f-contact-preferences/src/exceptions/index.js
@@ -1,0 +1,11 @@
+export class GetPreferencesError extends Error {
+    constructor (message, errorCode) {
+        super(message);
+        this.messageKey = 'errorMessages.pageLoad.description';
+        this.errorCode = errorCode;
+    }
+}
+
+export default {
+    GetPreferencesError
+};