+++ packages/components/pages/f-contact-preferences/src/tenants/en-GB.js
@@ -1,0 +1,27 @@
+const messages = {
+    errorMessages: {
+        errorHeading: 'Error',
+        pageLoad: {
+            description: 'Sorry, there was a problem fetching your contact preferences. Please try again later.'
+        }
+    },
+    heading: 'Contact Preferences',
+    locale: 'en-GB',
+    newsletter: {
+        email: 'Email',
+        sms: 'Text message',
+        subtitle: 'Get news and offers by'
+    },
+    orderUpdates: {
+        email: 'Email',
+        emailDescription: 'Order receipts are emailed to you',
+        sms: 'Text message',
+        smsDescription: 'For important updates only',
+        subtitle: 'Get order updates by'
+    },
+    saveChangesButton: 'Save changes'
+};
+
+export default {
+    messages
+};