+++ packages/components/pages/f-takeawaypay-activation/src/tenants/en-GB.js
@@ -1,0 +1,26 @@
+const messages = {
+    locale: 'en-GB',
+
+    messages: {
+        titleActivationAvailable: 'Welcome back, {consumerGivenName}. Let’s activate Just Eat Pay on your account.',
+        titleActivationNotLoggedIn: 'Let’s get you started with Just Eat Pay.',
+        titleActivationSuccessful: 'All systems go.',
+        titleActivationFailed: ' We can’t activate Just Eat Pay on your account right now, sorry.',
+        descriptionActivationAvailable1: 'Just Eat Pay will be activated for',
+        descriptionActivationAvailable2: 'Not you? Choose a different account or sign up.',
+        descriptionActivationNotLoggedIn: 'Do you already have a Just Eat account?',
+        descriptionActivationFailed: 'Try again or get in touch with your Just Eat Pay administrator for help.',
+        descriptionActivationSuccessful: 'Just Eat Pay is now activated on your account.'
+    },
+    actions: {
+        activateTakeawayPay: 'Activate Just Eat Pay',
+        startOrdering: 'Start ordering',
+        createAccount: 'Create a new account',
+        loginToAccount1: 'Use a different account',
+        loginToAccount2: 'I have an account'
+    }
+};
+
+export default {
+    messages
+};