+++ packages/components/pages/f-takeawaypay-activation/src/tenants/index.js
@@ -1,0 +1,9 @@
+import uk from './en-GB';
+import au from './en-AU';
+
+const tenantConfigs = {
+    'en-GB': uk,
+    'en-AU': au
+};
+
+export default tenantConfigs;